Update settings user preference pages
