+++ packages/frontend/app/(application)/settings/user-preferences/dashboard-settings/page.tsx
@@ -1,18 +1,13 @@
 import { Metadata } from 'next';
 
-import { DashboardSettingsContainer } from '$container/user-preferences/dashboard-settings.container';
-import { Layout } from '$layouts/Layout';
+import { DashboardSettingsContainer } from '$container/user-preferences/DashboardSettingsContainer';
 
 export const metadata: Metadata = {
   title: 'Dashboard Settings',
 };
 
 const DashboardSettingsUserPreferencePage = () => {
-  return (
-    <Layout title="Dashboard Settings">
-      <DashboardSettingsContainer />
-    </Layout>
-  );
+  return <DashboardSettingsContainer />;
 };
 
 export default DashboardSettingsUserPreferencePage;
