+++ packages/frontend/app/(application)/settings/user-preferences/market-update-settings/page.tsx
@@ -1,18 +1,13 @@
 import { Metadata } from 'next';
 
-import { MarketUpdateSettingsContainer } from '$container/user-preferences/market-update-settings.container';
-import { Layout } from '$layouts/Layout';
+import { MarketUpdateSettingsContainer } from '$container/user-preferences/MarketUpdateSettingsContainer';
 
 export const metadata: Metadata = {
   title: 'Market Update Settings',
 };
 
 const MarketUpdateSettingsUserPreferencePage = () => {
-  return (
-    <Layout title="Market Update Settings">
-      <MarketUpdateSettingsContainer />
-    </Layout>
-  );
+  return <MarketUpdateSettingsContainer />;
 };
 
 export default MarketUpdateSettingsUserPreferencePage;
