+++ packages/frontend/app/(application)/settings/user-preferences/maximum-items-per-page/page.tsx
@@ -1,18 +1,13 @@
 import { Metadata } from 'next';
 
-import { MaximumItemsPerPageContainer } from '$container/user-preferences/maximum-items-per-page.container';
-import { Layout } from '$layouts/Layout';
+import { MaximumItemsPerPageSettingsContainer } from '$container/user-preferences/MaximumItemsPerPageSettingsContainer';
 
 export const metadata: Metadata = {
   title: 'Maximum Items Per Page',
 };
 
 const MaximumItemsPerPageUserPreferencePage = () => {
-  return (
-    <Layout title="Maximum Items Per Page">
-      <MaximumItemsPerPageContainer />
-    </Layout>
-  );
+  return <MaximumItemsPerPageSettingsContainer />;
 };
 
 export default MaximumItemsPerPageUserPreferencePage;
