+++ packages/frontend/app/(application)/settings/user-preferences/page.tsx
@@ -1,19 +1,14 @@
 import { Metadata } from 'next';
 import { FC } from 'react';
 
-import { Layout } from '$layouts/Layout';
-import { UserPreferences } from '$views/settings/user-preferences/user-preferences';
+import { UserPreferencesContainer } from '$container/settings/UserPreferencesContainer';
 
 export const metadata: Metadata = {
   title: 'User Preferences',
 };
 
 const UserPreferenceListingPage: FC = () => {
-  return (
-    <Layout title="User Preferences">
-      <UserPreferences />
-    </Layout>
-  );
+  return <UserPreferencesContainer />;
 };
 
 export default UserPreferenceListingPage;
