+++ packages/frontend/containers/settings/UserPreferencesContainer.tsx
@@ -3,44 +3,43 @@ import { FC } from 'react';
 import { List } from '$blocks/List';
 import { ProminentLink } from '$blocks/ProminentLink';
 import { settingsPaths } from '$constants/settings-paths';
-import { UpdatePageInfo } from '$renderers/seo/updatePageInfo';
+import { Layout } from '$layouts/Layout';
 
-export const UserPreferences: FC = () => {
+export const UserPreferencesContainer: FC = () => {
   return (
-    <>
-      <UpdatePageInfo backLink={settingsPaths.default} />
+    <Layout title="User Preferences" backLink={settingsPaths.default}>
       <List>
         <ProminentLink
           link={`${settingsPaths.userPreferences}/dashboard-settings`}
           icon={'HomeIcon'}
         >
-          Dashboard settings
+          Dashboard Settings
         </ProminentLink>
         <ProminentLink
           link={`${settingsPaths.userPreferences}/statistics-settings`}
           icon={'ChartBarIcon'}
         >
-          Statistics settings
+          Statistics Page Settings
         </ProminentLink>
         <ProminentLink
           link={`${settingsPaths.userPreferences}/default-account-settings`}
           icon={'Squares2X2Icon'}
         >
-          Default account settings
+          Default Account Settings
         </ProminentLink>
         <ProminentLink
           link={`${settingsPaths.userPreferences}/maximum-items-per-page`}
           icon={'RectangleStackIcon'}
         >
-          Max amount of items per page
+          Maximum Items Per Page Settings
         </ProminentLink>
         <ProminentLink
           link={`${settingsPaths.userPreferences}/market-update-settings`}
           icon={'ArrowTrendingUpIcon'}
         >
-          Market update settings
+          Market Update Settings
         </ProminentLink>
       </List>
-    </>
+    </Layout>
   );
 };
