+++ packages/frontend/src/actions/settings/handleDashboardSettingsUpdate.ts
@@ -1,0 +1,19 @@
+import { redirect, RedirectType } from 'next/navigation';
+
+import { AccountType } from '$api/generated/financerApi';
+import { settingsPaths } from '$constants/settings-paths';
+import { DefaultFormActionHandler } from '$hooks/useFinancerFormState';
+import { UserPreferenceService } from '$ssr/api/user-preference.service';
+
+export const handleDashboardSettingsUpdate: DefaultFormActionHandler = async (
+  prev,
+  formData,
+) => {
+  'use server';
+
+  await UserPreferenceService.updateDashboardSettings({
+    accountTypes: formData.getAll('accountTypes') as AccountType[],
+  });
+
+  redirect(settingsPaths.userPreferences, RedirectType.push);
+};
