+++ packages/frontend/src/components/views/user-preferences/UserDefaultMarketUpdateSettingsForm.tsx
@@ -12,22 +12,21 @@ import {
   DefaultFormActionHandler,
   useFinancerFormState,
 } from '$hooks/useFinancerFormState';
-import { UpdatePageInfo } from '$renderers/seo/updatePageInfo';
 import { UserDefaultMarketUpdateSettings as UserDefaultMarketUpdateSettingsType } from '$ssr/api/user-preference.service';
 
-export interface UserDefaultMarketUpdateSettingsFormFields {
+export type UserDefaultMarketUpdateSettingsFormFields = {
   transactionDescription: string;
   category: string;
-}
+};
 
-interface UserDefaultMarketUpdateSettingsProps {
+type UserDefaultMarketUpdateSettingsFormProps = {
   categories: TransactionCategoryDtoWithCategoryTree[];
   data?: UserDefaultMarketUpdateSettingsType;
   onSave: DefaultFormActionHandler;
-}
+};
 
-export const UserDefaultMarketUpdateSettings: FC<
-  UserDefaultMarketUpdateSettingsProps
+export const UserDefaultMarketUpdateSettingsForm: FC<
+  UserDefaultMarketUpdateSettingsFormProps
 > = ({ categories, data, onSave }) => {
   const action = useFinancerFormState(
     'user-default-market-update-settings',
@@ -40,32 +39,29 @@ export const UserDefaultMarketUpdateSettings: FC<
   }, [data, methods]);
 
   return (
-    <>
-      <UpdatePageInfo backLink={settingsPaths.userPreferences} />
-      <Form
-        methods={methods}
-        action={action}
-        submitLabel="Save"
-        formFooterBackLink={settingsPaths.userPreferences}
-      >
-        <div className="grid gap-y-4 gap-x-4 sm:grid-cols-2">
-          <Input id="transactionDescription" isRequired>
-            Transaction description
-          </Input>
-          <Select
-            id="category"
-            options={[{ name: 'None', id: '' }, ...categories].map(
-              ({ name, id }) => ({
-                label: name,
-                value: id,
-              }),
-            )}
-            isRequired
-          >
-            Category
-          </Select>
-        </div>
-      </Form>
-    </>
+    <Form
+      methods={methods}
+      action={action}
+      submitLabel="Save"
+      formFooterBackLink={settingsPaths.userPreferences}
+    >
+      <div className="grid gap-y-4 gap-x-4 sm:grid-cols-2">
+        <Input id="transactionDescription" isRequired>
+          Transaction description
+        </Input>
+        <Select
+          id="category"
+          options={[{ name: 'None', id: '' }, ...categories].map(
+            ({ name, id }) => ({
+              label: name,
+              value: id,
+            }),
+          )}
+          isRequired
+        >
+          Category
+        </Select>
+      </div>
+    </Form>
   );
 };
