+++ packages/frontend/tsconfig.path.json
@@ -9,10 +9,11 @@
       "$elements/*": ["./src/components/elements/*"],
       "$layouts/*": ["./src/components/layouts/*"],
       "$renderers/*": ["./src/components/renderers/*"],
+      "$views/*": ["./src/components/views/*"],
       "$pages/*": ["./src/components/pages/*"],
-      "$views/*": ["./src/views/*"],
       "$hooks/*": ["./src/hooks/*"],
       "$assets/*": ["./src/assets/*"],
+      "$actions/*": ["./src/actions/*"],
       "$utils/*": ["./src/utils/*"],
       "$constants/*": ["./src/constants/*"],
       "$context/*": ["./src/context/*"],
